apply observable on relation and search api

+++ server/middlewares/relation.js
@@ -10,22 +10,23 @@ obj.follow = async (req,res) => {
 		res.status(400).send({ "message" : "자신을 팔로우 할 수 없습니다." });
 	} else {
 		const current = { toId, fromId };
-		const deleted = await db.Follow.destroy({ where : current });
-		if( deleted ){
-			res.send({ "data" : false });
-		} else {
-			const [updated] = await db.Follow.update({ deletedAt : null },{ where : current, paranoid : false });
-			if( updated ){
-        noticeMws.makeNotice(req.user,'follow',updated.dataValues.id,userIds);
-				res.send({ "data" : true });
-			} else {
-				db.Follow.create(current)
-				.then( created => {
-          makeNotice(req.user,'follow',current.dataValues.id,userIds);
-					res.send({ "data" : true });
-				});
-			}
-		}
+		const follow = await db.Follow.findOne({ where : current, paranoid : false });
+    const following = !follow || follow.dataValues.deletedAt;
+
+    if( follow && !following ){
+		  await db.Follow.destroy({ where : current });
+    } else if( follow && following ){
+      await db.Follow.update({ deletedAt : null }, { where : current, paranoid : false });
+      noticeMws.makeNotice(req.user,'follow',follow.dataValues.id,toId);
+    } else {
+			const created = await db.Follow.create(current);
+      noticeMws.makeNotice(req.user,'follow',created.dataValues.id,toId);
+    }
+    req.user.followings += following?1:-1;
+	  res.send({ "data" : {
+      to : toId,
+      following
+    }});
 	}
 }
